avatar uploading/updating tested and good, file uploading, deleteing and renaming utils generic and can be utilized in different ways, some amount of decsriptions in the controllers, PRing the image branches now
