+++ server/views/dashboard/profile-button.ejs
@@ -3,7 +3,7 @@
 <div class="flex flex-row bg-cyan-500 rounded-md">
 
   <div class="flex justify-center">
-    <img src="/images/<%=img_avatar %>" alt="Avatar" class="bg-cyan-100 rounded-[50%] max-w-10 h-auto">
+    <img src="/images/avatars/<%=img_avatar %>" alt="Avatar" class="bg-cyan-100 rounded-[50%] max-w-10 h-auto">
   </div>
 
   <div id='openProfileDialog' class="flex flex-row w-32 justify-between cursor-pointer hover:bg-white">
@@ -54,7 +54,7 @@
 		</p>
 	
 		
-		<img src="/images/<%=img_avatar %>" alt="Avatar" class="bg-cyan-100 rounded-[50%] max-w-10 h-auto">
+		<img src="/images/avatars/<%=img_avatar %>" alt="Avatar" class="bg-cyan-100 rounded-[50%] max-w-10 h-auto">
 		<i id="openAvatarInput" class='fas fa-pen text-xs flex'></i>
 		<div id="avatarInput" class="hidden">
 		  <form hx-post="/editAvatar/<%=username %>" enctype="multipart/form-data" hx-target="#dashboard-profile-button" hx-swap="innerHTML">
